Rename `GraphFacotryController` to `SystemController`
