+++ grakn-graph/src/main/java/ai/grakn/factory/GraknSessionImpl.java
@@ -36,7 +36,7 @@ import java.util.Properties;
 
 import static ai.grakn.util.EngineCommunicator.contactEngine;
 import static ai.grakn.util.REST.Request.GRAPH_CONFIG_PARAM;
-import static ai.grakn.util.REST.WebPath.Graph.CONFIGURATION;
+import static ai.grakn.util.REST.WebPath.System.CONFIGURATION;
 import static mjson.Json.read;
 
 /**
